Move ViewModels to appropriate namespace
