+++ src/Promenade.Web/Views/Home/Feature.cshtml
@@ -1,4 +1,4 @@
-﻿@model Ocuda.Promenade.Controllers.ViewModels.Locations.LocationDetailViewModel
+﻿@model Ocuda.Promenade.Controllers.ViewModels.Home.LocationDetailViewModel
 
 <div>
     <table width="100%" cellpadding="0" cellspacing="0" border="0">
